requirements.txt
mlflow==3.4.0
catboost==1.2.8
cloudpickle==3.1.1
numpy==2.1.3
pandas==2.3.2
scikit-learn==1.7.1
scipy==1.15.3
sagemaker-mlflow
psutil
pynvml
